refactor(fs): trim internal utilities
