+++ fs/_is_subdir_test.ts
@@ -1,0 +1,31 @@
+// Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
+// Copyright the Browserify authors. MIT License.
+
+import { assertEquals } from "../assert/mod.ts";
+import * as path from "../path/mod.ts";
+import { isSubdir } from "./_is_subdir.ts";
+
+Deno.test("isSubdir() returns a boolean indicating if dir is a subdir", function () {
+  const pairs = [
+    ["", "", false, path.posix.sep],
+    ["/first/second", "/first", false, path.posix.sep],
+    ["/first", "/first", false, path.posix.sep],
+    ["/first", "/first/second", true, path.posix.sep],
+    ["first", "first/second", true, path.posix.sep],
+    ["../first", "../first/second", true, path.posix.sep],
+    ["c:\\first", "c:\\first", false, path.win32.sep],
+    ["c:\\first", "c:\\first\\second", true, path.win32.sep],
+  ];
+
+  pairs.forEach(function (p) {
+    const src = p[0] as string;
+    const dest = p[1] as string;
+    const expected = p[2] as boolean;
+    const sep = p[3] as string;
+    assertEquals(
+      isSubdir(src, dest, sep),
+      expected,
+      `'${src}' should ${expected ? "" : "not"} be parent dir of '${dest}'`,
+    );
+  });
+});
