+++ fs/ensure_link.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 import { dirname } from "../path/dirname.ts";
 import { ensureDir, ensureDirSync } from "./ensure_dir.ts";
-import { toPathString } from "./_util.ts";
+import { toPathString } from "./_to_path_string.ts";
 
 /**
  * Ensures that the hard link exists.
