+++ fs/ensure_symlink.ts
@@ -2,7 +2,8 @@
 import { dirname } from "../path/dirname.ts";
 import { resolve } from "../path/resolve.ts";
 import { ensureDir, ensureDirSync } from "./ensure_dir.ts";
-import { getFileInfoType, toPathString } from "./_util.ts";
+import { getFileInfoType } from "./_get_file_info_type.ts";
+import { toPathString } from "./_to_path_string.ts";
 
 const isWindows = Deno.build.os === "windows";
